moved categories methods from admin-methods to categories.methods
