+++ imports/api/categories/categories.methods.js
@@ -1,0 +1,85 @@
+import { Meteor } from 'meteor/meteor';
+import { getHost } from '../@/shared';
+import Categories from './category';
+
+const catColors = [
+  'hsla(10, 62%, 80%, 0.7)',
+  'hsla(46, 62%, 80%, 0.7)',
+  'hsla(82, 62%, 80%, 0.7)',
+  'hsla(118, 62%, 80%, 0.7)',
+  'hsla(154, 62%, 80%, 0.7)',
+  'hsla(190, 62%, 80%, 0.7)',
+  'hsla(226, 62%, 80%, 0.7)',
+  'hsla(262, 62%, 80%, 0.7)',
+  'hsla(298, 62%, 80%, 0.7)',
+  'hsla(334, 62%, 80%, 0.7)',
+];
+
+const isUserAdmin = (members, userId) => {
+  return members.some(
+    (member) => member.id === userId && member.role === 'admin'
+  );
+};
+
+
+Meteor.methods({
+  getCategories() {
+    const user = Meteor.user();
+    const host = getHost(this);
+    if (!user) {
+      throw new Meteor.Error('You are not allowed');
+    }
+
+    return Categories.find({
+      host,
+    }).fetch();
+  },
+
+  addNewCategory(category, type) {
+    const user = Meteor.user();
+    const host = getHost(this);
+    const currentHost = Hosts.findOne({ host });
+    const isAdmin = currentHost && isUserAdmin(currentHost.members, user._id);
+
+    if (!user.isSuperAdmin && !isAdmin) {
+      throw new Meteor.Error('You are not allowed');
+    }
+
+    if (Categories.findOne({ label: category.toLowerCase(), host })) {
+      throw new Meteor.Error('Category already exists!');
+    }
+
+    const catLength = Categories.find({ host, type }).count();
+
+    try {
+      return Categories.insert({
+        host,
+        type,
+        label: category.toLowerCase(),
+        color: catColors[catLength],
+        addedBy: user._id,
+        addedUsername: user.username,
+        addedDate: new Date(),
+      });
+    } catch (error) {
+      throw new Meteor.Error(error);
+    }
+  },
+
+  removeCategory(categoryId) {
+    const user = Meteor.user();
+    const host = getHost(this);
+    const currentHost = Hosts.findOne({ host });
+    const isAdmin = currentHost && isUserAdmin(currentHost.members, user._id);
+
+    if (!user.isSuperAdmin && !isAdmin) {
+      throw new Meteor.Error('You are not allowed');
+    }
+
+    try {
+      Categories.remove(categoryId);
+    } catch (error) {
+      throw new Meteor.Error(error);
+    }
+  },
+});
